Create authentication, profile and home pages + refactor the API
Vue: Create the authentication system with the refresh token and a "remember me" button. The user can login, signup and logout from the application. If he is connected, he also can access to his profile page. The user session is managed with Vuex and also with the localStorage.

Before each routes, verify if the user is logged in and if he got the proper rights to access the page he wants.

Add an axios interceptor to add the Bearer token automatically to requests, and to intercept responses: if the response status is 401, try to reconnect the user automatically with the refresh token.

API: Add a "remember me" possibility, which return or not the refresh token on login